requirements.txt
asgiref==3.8.1
Django==5.0.7
django-rest-framework==0.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
sqlparse==0.5.1
